moving stats functions out of tsdf to their own module
