+++ python/tests/tsdf_tests.py
@@ -872,214 +872,6 @@ class TSDFBaseTests(SparkTest):
         ...
 
 
-class FourierTransformTest(SparkTest):
-    def test_fourier_transform(self):
-        """Test of fourier transform functionality in TSDF objects"""
-
-        # construct dataframes
-        tsdf_init = self.get_data_as_tsdf("init")
-        dfExpected = self.get_data_as_sdf("expected")
-
-        # convert to TSDF
-        result_tsdf = tsdf_init.fourier_transform(1, "val")
-
-        # should be equal to the expected dataframe
-        self.assertDataFrameEquality(result_tsdf.df, dfExpected)
-
-    def test_fourier_transform_valid_sequence_col_empty_partition_cols(self):
-        """Test of fourier transform functionality in TSDF objects"""
-
-        # construct dataframes
-        tsdf_init = self.get_data_as_tsdf("init")
-        dfExpected = self.get_data_as_sdf("expected")
-
-        # convert to TSDF
-        result_tsdf = tsdf_init.fourier_transform(1, "val")
-
-        # should be equal to the expected dataframe
-        self.assertDataFrameEquality(result_tsdf.df, dfExpected)
-
-    def test_fourier_transform_valid_sequence_col_valid_partition_cols(self):
-        """Test of fourier transform functionality in TSDF objects"""
-
-        # construct dataframes
-        tsdf_init = self.get_data_as_tsdf("init")
-        dfExpected = self.get_data_as_sdf("expected")
-
-        # convert to TSDF
-        result_tsdf = tsdf_init.fourier_transform(1, "val")
-
-        # should be equal to the expected dataframe
-        self.assertDataFrameEquality(result_tsdf.df, dfExpected)
-
-    def test_fourier_transform_no_sequence_col_empty_partition_cols(self):
-        """Test of fourier transform functionality in TSDF objects"""
-
-        # construct dataframes
-        tsdf_init = self.get_data_as_tsdf("init")
-        dfExpected = self.get_data_as_sdf("expected")
-
-        # convert to TSDF
-        result_tsdf = tsdf_init.fourier_transform(1, "val")
-
-        # should be equal to the expected dataframe
-        self.assertDataFrameEquality(result_tsdf.df, dfExpected)
-
-
-class RangeStatsTest(SparkTest):
-    def test_range_stats(self):
-        """Test of range stats for 20 minute rolling window"""
-
-        # construct dataframes
-        tsdf_init = self.get_data_as_tsdf("init")
-        dfExpected = self.get_data_as_sdf("expected")
-
-        # convert to TSDF
-
-        # using lookback of 20 minutes
-        featured_df = tsdf_init.withRangeStats(rangeBackWindowSecs=1200).df
-
-        # cast to decimal with precision in cents for simplicity
-        featured_df = featured_df.select(
-            sfn.col("symbol"),
-            sfn.col("event_ts"),
-            sfn.col("mean_trade_pr").cast("decimal(5, 2)"),
-            sfn.col("count_trade_pr"),
-            sfn.col("min_trade_pr").cast("decimal(5,2)"),
-            sfn.col("max_trade_pr").cast("decimal(5,2)"),
-            sfn.col("sum_trade_pr").cast("decimal(5,2)"),
-            sfn.col("stddev_trade_pr").cast("decimal(5,2)"),
-            sfn.col("zscore_trade_pr").cast("decimal(5,2)"),
-        )
-
-        # cast to decimal with precision in cents for simplicity
-        dfExpected = dfExpected.select(
-            sfn.col("symbol"),
-            sfn.col("event_ts"),
-            sfn.col("mean_trade_pr").cast("decimal(5, 2)"),
-            sfn.col("count_trade_pr"),
-            sfn.col("min_trade_pr").cast("decimal(5,2)"),
-            sfn.col("max_trade_pr").cast("decimal(5,2)"),
-            sfn.col("sum_trade_pr").cast("decimal(5,2)"),
-            sfn.col("stddev_trade_pr").cast("decimal(5,2)"),
-            sfn.col("zscore_trade_pr").cast("decimal(5,2)"),
-        )
-
-        # should be equal to the expected dataframe
-        self.assertDataFrameEquality(featured_df, dfExpected)
-
-    def test_group_stats(self):
-        """Test of range stats for 20 minute rolling window"""
-
-        # construct dataframes
-        tsdf_init = self.get_data_as_tsdf("init")
-        dfExpected = self.get_data_as_sdf("expected")
-
-        # using lookback of 20 minutes
-        featured_df = tsdf_init.withGroupedStats(freq="1 min").df
-
-        # cast to decimal with precision in cents for simplicity
-        featured_df = featured_df.select(
-            sfn.col("symbol"),
-            sfn.col("event_ts"),
-            sfn.col("mean_trade_pr").cast("decimal(5, 2)"),
-            sfn.col("count_trade_pr"),
-            sfn.col("min_trade_pr").cast("decimal(5,2)"),
-            sfn.col("max_trade_pr").cast("decimal(5,2)"),
-            sfn.col("sum_trade_pr").cast("decimal(5,2)"),
-            sfn.col("stddev_trade_pr").cast("decimal(5,2)"),
-        )
-
-        # cast to decimal with precision in cents for simplicity
-        dfExpected = dfExpected.select(
-            sfn.col("symbol"),
-            sfn.col("event_ts"),
-            sfn.col("mean_trade_pr").cast("decimal(5, 2)"),
-            sfn.col("count_trade_pr"),
-            sfn.col("min_trade_pr").cast("decimal(5,2)"),
-            sfn.col("max_trade_pr").cast("decimal(5,2)"),
-            sfn.col("sum_trade_pr").cast("decimal(5,2)"),
-            sfn.col("stddev_trade_pr").cast("decimal(5,2)"),
-        )
-
-        # should be equal to the expected dataframe
-        self.assertDataFrameEquality(featured_df, dfExpected)
-
-
-class ResampleTest(SparkTest):
-    def test_resample(self):
-        """Test of range stats for 20 minute rolling window"""
-
-        # construct dataframes
-        tsdf_input = self.get_data_as_tsdf("input")
-        dfExpected = self.get_data_as_sdf("expected")
-        expected_30s_df = self.get_data_as_sdf("expected30m")
-        barsExpected = self.get_data_as_sdf("expectedbars")
-
-        # 1 minute aggregation
-        featured_df = tsdf_input.resample(freq="min", func="floor", prefix="floor").df
-        # 30 minute aggregation
-        resample_30m = tsdf_input.resample(freq="5 minutes", func="mean").df.withColumn(
-            "trade_pr", sfn.round(sfn.col("trade_pr"), 2)
-        )
-
-        bars = tsdf_input.calc_bars(
-            freq="min", metricCols=["trade_pr", "trade_pr_2"]
-        ).df
-
-        # should be equal to the expected dataframe
-        self.assertDataFrameEquality(featured_df, dfExpected)
-        self.assertDataFrameEquality(resample_30m, expected_30s_df)
-
-        # test bars summary
-        self.assertDataFrameEquality(bars, barsExpected)
-
-    def test_resample_millis(self):
-        """Test of resampling for millisecond windows"""
-
-        # construct dataframes
-        tsdf_init = self.get_data_as_tsdf("init")
-        dfExpected = self.get_data_as_sdf("expectedms")
-
-        # 30 minute aggregation
-        resample_ms = tsdf_init.resample(freq="ms", func="mean").df.withColumn(
-            "trade_pr", sfn.round(sfn.col("trade_pr"), 2)
-        )
-
-        self.assertDataFrameEquality(resample_ms, dfExpected)
-
-    def test_upsample(self):
-        """Test of range stats for 20 minute rolling window"""
-
-        # construct dataframes
-        tsdf_input = self.get_data_as_tsdf("input")
-        expected_30s_df = self.get_data_as_sdf("expected30m")
-        barsExpected = self.get_data_as_sdf("expectedbars")
-
-        resample_30m = tsdf_input.resample(
-            freq="5 minutes", func="mean", fill=True
-        ).df.withColumn("trade_pr", sfn.round(sfn.col("trade_pr"), 2))
-
-        bars = tsdf_input.calc_bars(
-            freq="min", metricCols=["trade_pr", "trade_pr_2"]
-        ).df
-
-        upsampled = resample_30m.filter(
-            sfn.col("event_ts").isin(
-                "2020-08-01 00:00:00",
-                "2020-08-01 00:05:00",
-                "2020-09-01 00:00:00",
-                "2020-09-01 00:15:00",
-            )
-        )
-
-        # test upsample summary
-        self.assertDataFrameEquality(upsampled, expected_30s_df)
-
-        # test bars summary
-        self.assertDataFrameEquality(bars, barsExpected)
-
-
 class ExtractStateIntervalsTest(SparkTest):
     """Test of finding time ranges for metrics with constant state."""
